requirements.txt
Django==3.0.8
djangorestframework==3.11.0
dj-database-url==0.2.1
psycopg2-binary==2.8.2
gunicorn==20.0.4
whitenoise==5.2.0
coverage==5.4
flake8==3.8.4
